+++ src/Context.cpp
@@ -1698,9 +1698,32 @@ int MGLContext_set_wireframe(MGLContext * self, PyObject * value) {
 		self->gl.PolygonMode(GL_FRONT_AND_BACK, GL_FILL);
 		self->wireframe = false;
 	} else {
-		MGLError_Set("invalid value for enabled");
+		MGLError_Set("invalid value for wireframe");
 		return -1;
 	}
+	return 0;
+}
+
+PyObject * MGLContext_get_front_face(MGLContext * self) {
+	if (self->front_face == GL_CW) {
+		return PyUnicode_FromString("CW");
+	}
+	return PyUnicode_FromString("CCW");
+}
+
+int MGLContext_set_front_face(MGLContext * self, PyObject * value) {
+	const char * str = PyUnicode_AsUTF8(value);
+
+	if (!strcmp(str, "CW")) {
+		self->front_face = GL_CW;
+	} else if (!strcmp(str, "CCW")) {
+		self->front_face = GL_CCW;
+	} else {
+		MGLError_Set("invalid value for front_face");
+		return -1;
+	}
+
+	self->gl.FrontFace(self->front_face);
 	return 0;
 }
 
@@ -1988,9 +2011,6 @@ PyObject * MGLContext_get_info(MGLContext * self, void * closure) {
 		int gl_max_server_wait_timeout = 0;
 		gl.GetIntegerv(GL_MAX_SERVER_WAIT_TIMEOUT, &gl_max_server_wait_timeout);
 
-		int gl_max_shader_storage_buffer_bindings = 0;
-		gl.GetIntegerv(GL_MAX_SHADER_STORAGE_BUFFER_BINDINGS, &gl_max_shader_storage_buffer_bindings);
-
 		int gl_max_texture_buffer_size = 0;
 		gl.GetIntegerv(GL_MAX_TEXTURE_BUFFER_SIZE, &gl_max_texture_buffer_size);
 
@@ -2072,7 +2092,6 @@ PyObject * MGLContext_get_info(MGLContext * self, void * closure) {
 		PyDict_SetItemString(info, "GL_MAX_RENDERBUFFER_SIZE", PyLong_FromLong(gl_max_renderbuffer_size));
 		PyDict_SetItemString(info, "GL_MAX_SAMPLE_MASK_WORDS", PyLong_FromLong(gl_max_sample_mask_words));
 		PyDict_SetItemString(info, "GL_MAX_SERVER_WAIT_TIMEOUT", PyLong_FromLong(gl_max_server_wait_timeout));
-		PyDict_SetItemString(info, "GL_MAX_SHADER_STORAGE_BUFFER_BINDINGS", PyLong_FromLong(gl_max_shader_storage_buffer_bindings));
 		PyDict_SetItemString(info, "GL_MAX_TEXTURE_BUFFER_SIZE", PyLong_FromLong(gl_max_texture_buffer_size));
 		PyDict_SetItemString(info, "GL_MAX_TEXTURE_IMAGE_UNITS", PyLong_FromLong(gl_max_texture_image_units));
 		PyDict_SetItemString(info, "GL_MAX_TEXTURE_LOD_BIAS", PyLong_FromLong(gl_max_texture_lod_bias));
@@ -2180,6 +2199,9 @@ PyObject * MGLContext_get_info(MGLContext * self, void * closure) {
 			)
 		);
 
+		int gl_max_shader_storage_buffer_bindings = 0;
+		gl.GetIntegerv(GL_MAX_SHADER_STORAGE_BUFFER_BINDINGS, &gl_max_shader_storage_buffer_bindings);
+
 		int gl_max_combined_shader_storage_blocks = 0;
 		gl.GetIntegerv(GL_MAX_COMBINED_SHADER_STORAGE_BLOCKS, &gl_max_combined_shader_storage_blocks);
 
@@ -2240,6 +2262,7 @@ PyObject * MGLContext_get_info(MGLContext * self, void * closure) {
 		int gl_max_element_index = 0;
 		gl.GetIntegerv(GL_MAX_ELEMENT_INDEX, &gl_max_element_index);
 
+		PyDict_SetItemString(info, "GL_MAX_SHADER_STORAGE_BUFFER_BINDINGS", PyLong_FromLong(gl_max_shader_storage_buffer_bindings));
 		PyDict_SetItemString(info, "GL_MAX_COMBINED_SHADER_STORAGE_BLOCKS", PyLong_FromLong(gl_max_combined_shader_storage_blocks));
 		PyDict_SetItemString(info, "GL_MAX_VERTEX_SHADER_STORAGE_BLOCKS", PyLong_FromLong(gl_max_vertex_shader_storage_blocks));
 		PyDict_SetItemString(info, "GL_MAX_FRAGMENT_SHADER_STORAGE_BLOCKS", PyLong_FromLong(gl_max_fragment_shader_storage_blocks));
@@ -2285,6 +2308,7 @@ PyGetSetDef MGLContext_tp_getseters[] = {
 	{(char *)"default_framebuffer", (getter)MGLContext_get_default_framebuffer, 0, 0, 0},
 
 	{(char *)"wireframe", (getter)MGLContext_get_wireframe, (setter)MGLContext_set_wireframe, 0, 0},
+	{(char *)"front_face", (getter)MGLContext_get_front_face, (setter)MGLContext_set_front_face, 0, 0},
 
 	{(char *)"error", (getter)MGLContext_get_error, 0, 0, 0},
 	{(char *)"vendor", (getter)MGLContext_get_vendor, 0, 0, 0},
@@ -2437,4 +2461,7 @@ void MGLContext_Initialize(MGLContext * self) {
 
 	Py_INCREF(self->default_framebuffer);
 	self->bound_framebuffer = self->default_framebuffer;
+
+	self->wireframe = false;
+	self->front_face = GL_CCW;
 }